streamlit-app/requirements.txt
streamlit
openai
toml
requests
google-cloud-firestore
pandas
python-dotenv
